Revert "Moved some files to Components package"
This reverts commit a5834cd3028673bfaf0c1a6ede2f1146ae668e58.
