+++ app/src/main/java/com/belia/speedotransfer/ui/common_ui/TopBar.kt
@@ -1,9 +1,10 @@
-package com.belia.speedotransfer.ui.components
+package com.belia.speedotransfer.ui.common_ui
 
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
@@ -11,8 +12,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.sp
 import com.belia.speedotransfer.R
-import com.belia.speedotransfer.ui.theme.titleMedium
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -26,12 +27,13 @@ fun TopBar(
         title = {
             Text(
                 text = title,
-                style = titleMedium,
+                style = MaterialTheme.typography.titleMedium,
+                fontSize = 20.sp,
             )
         },
         navigationIcon = {
             if (hasIcon) {
-                IconButton(onClick = { /*TODO(popBackStack)*/ }) {
+                IconButton(onClick = { /*TODO*/ }) {
                     Icon(
                         painter = painterResource(id = R.drawable.ic_back),
                         contentDescription = "back button"
@@ -43,4 +45,10 @@ fun TopBar(
             containerColor = color
         ),
     )
+}
+
+@Preview
+@Composable
+private fun TopBarPrev() {
+    TopBar(color = Color(0xFFFFF8E7), hasIcon = true, title = "Transfer")
 }
